completed code along, creating components and passing props, .map in react
